pyproject.toml
[tool.poetry]
name = "sudoisytdl"
version = "0.1.0"
description = ""
authors = ["Benedikt Kristinsson <benedikt@lokun.is>"]

[tool.poetry.dependencies]
python = "^3.9"
loguru = "^0.5.3"
python-telegram-bot = "^13.0"
youtube-dl = "^2020.11.19"
ffmpeg-python = "^0.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
ytdl = 'sudoisytdl.cli:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
